add rab id in migration cashflow

+++ database/migrations/2020_03_17_132745_create_cash_flows_table.php
@@ -15,7 +15,7 @@ class CreateCashFlowsTable extends Migration
     {
         Schema::create('cash_flows', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('rab_id')->nullable();
+            $table->unsignedBigInteger('rab_id')->nullable();
             $table->unsignedInteger('branch_id');
             $table->date('date');
             $table->string('month', 10);
